checked all style
