+++ src/main/java/core/model/Dog.java
@@ -1,0 +1,4 @@
+package core.model;
+
+public class Dog {
+}
