+++ src/test/java/core/HelloWorldTest.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package core;
 
 /**
  * Feel free to remove this class and create your own.
